+++ source/plugins/studio/system.browser/index.js
@@ -1,43 +1,58 @@
 import { EventEmitter } from 'events';
 
-let system = {
-
-	events: new EventEmitter(),
-
-	close ()
-	{
-		location.href = 'https://wyliodrin.studio';
-	},
-
-	minimize ()
-	{
-		
-	},
-
-	fullscreen ()
-	{
-		let isFullscreen = () => {
-			return document.fullscreenElement !== null;
-		};
-		if (isFullscreen ())
-		{
-			document.exitFullscreen ();
-		}
-		else
-		{
-			document.body.requestFullscreen ();
-		}
-	},
-	openLink (url)
-	{
-		window.open(url);
-	},
-	platform ()
-	{
-		return 'browser';
-	}
-};
-
 export function setup (options, imports, register) {
+	let serviceWorker = null;
+
+	let system = {
+
+		events: new EventEmitter(),
+	
+		send (tag, data)
+		{
+			serviceWorker.postMessage({
+				tag,
+				data
+			});
+		},
+	
+		close ()
+		{
+			location.href = 'https://wyliodrin.studio';
+		},
+	
+		minimize ()
+		{
+			
+		},
+	
+		fullscreen ()
+		{
+			let isFullscreen = () => {
+				return document.fullscreenElement !== null;
+			};
+			if (isFullscreen ())
+			{
+				document.exitFullscreen ();
+			}
+			else
+			{
+				document.body.requestFullscreen ();
+			}
+		},
+		openLink (url)
+		{
+			window.open(url);
+		},
+		platform ()
+		{
+			return 'browser';
+		}
+	};
+
+	navigator.serviceWorker.ready.then( registration => {
+		serviceWorker = registration.active;
+	});
+	
+
 	register (null, { system });
 }