+++ source/plugins/studio/system.electron/index.js
@@ -7,6 +7,11 @@ let system = {
 
 	events: new EventEmitter (),
 	
+	send (tag, data)
+	{
+		ipcRenderer.send (tag, data);
+	},
+
 	close ()
 	{
 		ipcRenderer.send ('close');
@@ -53,8 +58,8 @@ export function setup (options, imports, register) {
 		system.events.emit ('close-ask');
 	});
 
-	ipcRenderer.on('updated', () => {
-		system.events.emit('updated');
+	ipcRenderer.on ('update-ask', () => {
+		system.events.emit ('update-ask');
 	});
 
 	register (null, { system });