+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Wyliodrin STUDIO 2.3.0
+  * **added** serialport plugin, works in electron and Chrome browser
   * **fixed** schematics and enabled it for browser
 ## Wyliodrin STUDIO 2.2.1-beta
   * **new** Raspberry Pi images (with docker)