+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Wyliodrin STUDIO 2.3.0
   * **added** serialport plugin, works in electron and Chrome browser
+  * **added** update plugin, handles updates in electron and browser
   * **fixed** schematics and enbled it for browser
   * **fixed** display micropython options after port selection in browser
   * **fixed** schematics image layout